requirements.txt
attrs==20.3.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.15.8
pycryptodome==3.10.1
pyrsistent==0.17.3
python-dotenv==0.17.1
requests==2.25.1
rlp==2.0.1
six==1.15.0
toolz==0.11.1
urllib3==1.26.4
varint==1.0.2
web3==5.19.0
websockets==8.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
decorator==5.0.7
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.17.2
jsonschema==3.2.0
lru-dict==1.1.7
matplotlib-inline==0.1.2
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pycryptodome==3.10.1
Pygments==2.8.1
pyrsistent==0.17.3
python-dotenv==0.17.1
requests==2.25.1
rlp==2.0.1
six==1.15.0
toolz==0.11.1
traitlets==5.0.5
urllib3==1.26.4
varint==1.0.2
wcwidth==0.2.5
web3==5.19.0
websockets==8.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
decorator==5.0.7
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-tester==0.5.0b4
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.17.2
jsonschema==3.2.0
lru-dict==1.1.7
matplotlib-inline==0.1.2
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pycryptodome==3.10.1
Pygments==2.8.1
pyrsistent==0.17.3
python-dotenv==0.17.1
requests==2.25.1
rlp==2.0.1
semantic-version==2.8.5
six==1.15.0
toolz==0.11.1
traitlets==5.0.5
urllib3==1.26.4
varint==1.0.2
wcwidth==0.2.5
web3==5.19.0
websockets==8.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
decorator==5.0.7
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-tester==0.5.0b4
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.17.2
jsonschema==3.2.0
lru-dict==1.1.7
matplotlib-inline==0.1.2
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pycryptodome==3.10.1
Pygments==2.8.1
pyrsistent==0.17.3
python-dotenv==0.17.1
requests==2.25.1
rlp==2.0.1
semantic-version==2.8.5
six==1.15.0
toolz==0.11.1
traitlets==5.0.5
urllib3==1.26.4
varint==1.0.2
wcwidth==0.2.5
web3==5.19.0
websockets==8.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
decorator==5.0.7
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-tester==0.5.0b4
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.17.2
jsonschema==3.2.0
lru-dict==1.1.7
matplotlib-inline==0.1.2
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
py-solc-x==1.1.0
pycryptodome==3.10.1
Pygments==2.8.1
pyrsistent==0.17.3
python-dotenv==0.17.1
requests==2.25.1
rlp==2.0.1
semantic-version==2.8.5
six==1.15.0
toolz==0.11.1
traitlets==5.0.5
urllib3==1.26.4
varint==1.0.2
wcwidth==0.2.5
web3==5.19.0
websockets==8.1
